+++ lib/src/hashlib_base.dart
@@ -1,1 +1,2 @@
 export 'package:hashlib/src/md5.dart';
+export 'package:hashlib/src/sha1.dart';